refactor(canister-core): split serving and access protection out of state.rs
After the Store and Certifier extractions, state.rs still carried the whole
HTTP read path and the entire access-protection surface alongside the
orchestration core. Move them into two sibling `impl State` files, matching the
existing sync.rs precedent:

- serving.rs: http_request + the gate dispatch, matching_rule, and the build_*
  response helpers (the read path — reads Store for content, Certifier for
  witnesses, mutates nothing).
- access.rs: the protection serve helpers (serve_unauthenticated / serve_redeem
  / serve_protection_response / cookie_token_valid) and the controller-driven
  state machine (enable/disable/issue/revoke/list/check + token GC).

state.rs is now the orchestration core: struct + construction, asset CRUD,
queries, the state-hash/manifest fold, redirect-rule delegators, env cookie,
and upgrade rebuild — 563 lines, down from 1746 before this branch.

store/certifier/token_index become pub(crate) so the split impls can reach
them (like chunks/sync_session already were for sync.rs); the composed
sub-objects stay fully encapsulated. Pure code motion — no behavior change.

Verified: builds (incl. canbench-rs), 158 unit tests pass, fmt + clippy clean,
full workspace builds, e2e 26/26 pass.
